fix: EZ Suite dict->dataclass, dialog double-click, cancel button, shortcuts, PSF consistency
- EZ Suite: all 10 steps now guard isinstance(params, dict) and construct the correct dataclass. Removed invalid _osc_params(). Fixed _stretch_params() black_clip->black_point. histogram step uses defaults (not stretch_factor). abe step unpacks tuple. color_calibrate extracts .data from dataclass.
- EZ Script dialog: _run() now calls accept() so it closes after OK + disables button box during processing to prevent double-click. Renamed result() to dialog_result() to avoid shadowing QDialog.result().
- Cancel button: now calls QThread.requestInterruption(); stale workers are cancelled/waited/terminated instead of refusing new work. All terminal signals reset self._worker = None.
- Keyboard shortcuts (Ctrl+H/B/Shift+P): removed third shortcut arg from _tbtn() to prevent duplicate QShortcut instances eating key events.
- Statistics/FITS Header buttons: log warning when no image loaded instead of silent no-op. FITS Header had wrong import (header_dialog -> fits_header_dialog).
- PSF measurement: always uses CPU detection (OpenCV + roundness) for deterministic FWHM regardless of force_cpu flag. Removed unused torch imports.
- CosmicClarity model URLs: use real release tag (Windows) and actual model filenames. Incompatible state dicts log warning and return None gracefully.
- Processing Graph: QTimer polling every 1s for auto-refresh.
- Ctrl+Shift+B: keyPressEvent checks event.modifiers()==NoModifier.
- Ctrl+H ambiguous: FITS Header moved to Ctrl+Shift+H.
- CosmicClarity apply(): added progress kwarg.
- FITS Header button added to tools panel Utility tab.
- Preferences crash: has_gpu() -> is_gpu property.
- PixelMath, Blink, ChannelMatch, LensDistortion, LocalNormalization, MureDenoise, SuperBias, DBE batch preprocessing, Processing Graph, live stacking — all wired into menus/toolbar with working tests.
- installer.sh deleted (leftover from unrelated project).
- BatchPreprocessing wizard: 4-page PixInsight-style pipeline.
- astrometry.net: reverted to request-json= form-encoding. 429 handling. API key whitespace stripped on save/retrieve.
- .gitignore: node_modules/, package.json, package-lock.json, .opencode/

+++ cosmica/core/local_contrast.py
@@ -95,7 +95,7 @@ def _apply_clahe_mono(
     if amount < 1.0:
         enhanced_f = data * (1 - amount) + enhanced_f * amount
 
-    return np.clip(enhanced_f, 0, 1)
+    return np.clip(enhanced_f, 0, 1).astype(np.float32)
 
 
 def _apply_clahe_color(
